egfr-ai-drug-discovery/requirements.txt
# Core numerical & data
numpy==1.24.4
pandas==2.0.3
scipy==1.10.1

# Machine Learning / Deep Learning
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# Cheminformatics
rdkit-pypi==2022.9.5
selfies==2.1.1

# Visualization
matplotlib==3.7.5
seaborn==0.12.2

# Notebook environment
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.29.4

# Utilities
tqdm==4.66.4
pyyaml==6.0.1

# Optional (recommended for chemistry workflows)
scikit-learn==1.3.2

# Optional docking support (requires system install)
# NOTE: These are Python wrappers/utilities; core tools installed separately
openbabel-wheel==3.1.1.20

# File handling / misc
joblib==1.3.2